requirements.txt
altair==4.2.2
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
blinker==1.6.2
branca==0.6.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
entrypoints==0.4
executing==1.2.0
Fiona==1.9.3
folium==0.14.0
fonttools==4.39.3
geopandas==0.12.2
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
importlib-metadata==6.3.0
ipykernel==6.22.0
ipython==8.12.0
jedi==0.18.2
Jinja2==3.1.2
jsonschema==4.17.3
jupyter_client==8.2.0
jupyter_core==5.3.0
kiwisolver==1.4.4
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
munch==2.5.0
nest-asyncio==1.5.6
networkx==3.1
numpy==1.24.2
osmnx==1.3.0
packaging==23.1
pandas==1.5.3
parso==0.8.3
patsy==0.5.3
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.2.0
plotly==5.14.1
plotly-express==0.4.1
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.4
pure-eval==0.2.2
pyarrow==11.0.0
pydeck==0.8.0
Pygments==2.15.0
Pympler==1.0.1
pyparsing==3.0.9
pyproj==3.5.0
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pywin32==306
pyzmq==25.0.2
requests==2.28.2
rich==13.3.4
Rtree==1.0.1
scipy==1.10.1
shapely==2.0.1
six==1.16.0
smmap==5.0.0
stack-data==0.6.2
statsmodels==0.13.5
streamlit==1.21.0
streamlit-folium==0.11.1
streamlit-lottie==0.0.3
tenacity==8.2.2
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
validators==0.20.0
watchdog==3.0.0
wcwidth==0.2.6
zipp==3.15.0
